Move TreeNode definition
